Checkpoint work on path of building export
We can successfully export codes to Path of Building. However,
it soft-crashed the entire calculation portion of the program.

Looks like the export contains a bunch of internal state which
is wholesale jammed in. That makes sense as an internal format
but, oh buddy, I'm not investing the time to deal with that here.